More wip on offline puzzles
