+++ src/ui/training/index.ts
@@ -12,7 +12,7 @@ import { renderContent, renderHeader, overlay } from './trainingView'
 import * as xhr from './xhr'
 import TrainingCtrl from './TrainingCtrl'
 import { connectingHeader } from '../shared/common'
-import { syncAndLoadNewPuzzle, puzzleLoadFailure } from './utils'
+import { syncAndLoadNewPuzzle, puzzleLoadFailure } from './offlineService'
 import { PuzzleData } from '../../lichess/interfaces/training'
 import database from './database'
